Consolidate favorites route tests into one file
