+++ tests/favorites.spec.js
@@ -8,15 +8,16 @@ const database = require('knex')(configuration);
 
 
 describe('Test the favorites route', () => {
-  beforeEach(async () => {
-    await database.raw('truncate table favorites cascade');
-    fetch.resetMocks();
-  });
-  afterEach(() => {
-    database.raw('truncate table favorites cascade');
-  });
 
   describe('POST favorites', () => {
+    beforeEach(async () => {
+      await database.raw('truncate table favorites cascade');
+      fetch.resetMocks();
+    });
+    afterEach(() => {
+      database.raw('truncate table favorites cascade');
+    });
+
     test('It should respond to the POST method', async () => {
       let newFav = {
         title: "We will Rock You",
@@ -281,4 +282,183 @@ describe('Test the favorites route', () => {
       expect(res_2.body).toHaveProperty("detail", "That song is already favorited.")
     })
   })
+
+  describe('GET favorite by id', () => {
+    beforeEach(async () => {
+      await database.raw('truncate table favorites cascade');
+      let favorites = [{
+          title: 'Hound Dog',
+          artistName: 'Elvis Presley',
+          genre: 'Rock',
+          rating: 47
+        },
+        {
+          title: 'Umbrella',
+          artistName: 'Rihanna',
+          genre: 'Pop',
+          rating: 83
+        },
+        {
+          title: 'everything i wanted',
+          artistName: 'Billie Eilish',
+          genre: 'Music',
+          rating: 100
+        }
+      ];
+      await database('favorites').insert(favorites, 'id');
+    });
+    afterEach(() => {
+      database.raw('truncate table favorites cascade');
+    });
+
+    test('It should respond to the GET method', async() =>{
+      let favOne = await database('favorites').select('id').first()
+
+      const res = await request(app)
+        .get(`/api/v1/favorites/${favOne.id}`);
+
+      expect(res.statusCode).toBe(200);
+      expect(res.body).toHaveProperty("title", "Hound Dog");
+      expect(res.body).toHaveProperty("artistName", "Elvis Presley");
+      expect(res.body).toHaveProperty("genre", "Rock");
+      expect(res.body).toHaveProperty("rating", 47);
+    })
+
+    test('It should return a 404 if id not in favorites table', async() => {
+      let notAFav = -1
+
+      const res = await request(app)
+        .get(`/api/v1/favorites/${notAFav}`);
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toHaveProperty("error", "No such favorite found");
+    })
+
+    test('It should return a 500 if other error comes up', async() => {
+      let errorFav = 'abc'
+
+      const res = await request(app)
+        .get(`/api/v1/favorites/${errorFav}`);
+
+      expect(res.statusCode).toBe(500);
+      expect(res.body).toHaveProperty("error");
+    })
+  })
+
+  describe('DELETE favorite by id', () => {
+    beforeEach(async () => {
+      await database.raw('truncate table favorites cascade');
+      let favorites = [{
+        title: 'Hound Dog',
+        artistName: 'Elvis Presley',
+        genre: 'Rock',
+        rating: 47
+      },
+      {
+        title: 'Umbrella',
+        artistName: 'Rihanna',
+        genre: 'Pop',
+        rating: 83
+      },
+      {
+        title: 'everything i wanted',
+        artistName: 'Billie Eilish',
+        genre: 'Music',
+        rating: 100
+      }];
+      await database('favorites').insert(favorites, 'id');
+    });
+    afterEach(() => {
+      database.raw('truncate table favorites cascade');
+    });
+
+    test('It should respond to the DELETE method', async() =>{
+      let favsCount = await database('favorites').select('id').then((favs) => favs.length)
+      let favOne = await database('favorites').select('id').first()
+
+      const res = await request(app)
+        .delete(`/api/v1/favorites/${favOne.id}`);
+
+      expect(res.statusCode).toBe(204);
+
+      let newFavsCount = await database('favorites').select('id').then((favs) => favs.length)
+      expect(newFavsCount).toBe(favsCount - 1)
+
+      let favNotFound = await database('favorites').where(favOne).select()
+      expect(favNotFound.length).toBe(0)
+    })
+
+    test('It should return a 404 if id not in favorites table', async() => {
+      let notAFavID = -1
+
+      const res = await request(app)
+        .delete(`/api/v1/favorites/${notAFavID}`);
+
+      expect(res.statusCode).toBe(404);
+      expect(res.body).toHaveProperty("error", "No such favorite found. No deletion made.");
+    })
+
+    test('It should return a 500 if other error comes up', async() => {
+      let errorFav = 'abc'
+
+      const res = await request(app)
+        .delete(`/api/v1/favorites/${errorFav}`);
+
+      expect(res.statusCode).toBe(500);
+      expect(res.body).toHaveProperty("error");
+    })
+  })
+
+  describe('GET favorites', () => {
+    beforeEach(async () => {
+      await database.raw('truncate table favorites cascade');
+    });
+
+    afterEach(() => {
+      database.raw('truncate table favorites cascade');
+    });
+
+    test('It should respond with a list of favorites', async() => {
+      let favorites_array = [
+        {
+          "title": "We Will Rock You",
+          "artistName": "Queen",
+          "genre": "Rock",
+          "rating": 88
+        },
+        {
+          "title": "Careless Whisper",
+          "artistName": "George Michael",
+          "genre": "Pop",
+          "rating": 93
+        }
+      ]
+
+      await database('favorites').insert(favorites_array, 'id');
+
+      const res = await request(app)
+        .get("/api/v1/favorites")
+
+      expect(res.statusCode).toBe(200);
+      expect(res.body.data.length).toBe(2);
+
+      expect(res.body.data[0].title).toBe("We Will Rock You");
+      expect(res.body.data[0].artistName).toBe("Queen");
+      expect(res.body.data[0].genre).toBe("Rock");
+      expect(res.body.data[0].rating).toBe(88);
+
+      expect(res.body.data[1].title).toBe("Careless Whisper");
+      expect(res.body.data[1].artistName).toBe("George Michael");
+      expect(res.body.data[1].genre).toBe("Pop");
+      expect(res.body.data[1].rating).toBe(93);
+    });
+
+    test("It should respond with an empty array when no favorites present", async() => {
+      const res = await request(app)
+        .get("/api/v1/favorites")
+
+      expect(res.statusCode).toBe(200);
+      expect(res.body).toHaveProperty("data", []);
+    });
+  });
 });
